+++ anom.hpp
@@ -6,6 +6,8 @@
 
 using namespace std;
 
-vector<double>* mstream(vector<vector<double> >& numeric, vector<vector<int> >& categ, vector<int>& times, int num_rows, int num_buckets, double factor, int dimension1, int dimension2);
+vector<double> *
+mstream(vector<vector<double> > &numeric, vector<vector<long> > &categ, vector<int> &times, int num_rows,
+        int num_buckets, double factor, int dimension1, int dimension2);
 
 #endif /* anom_hpp */
